Added UI support for managing Syllable patterns
