+++ langforge-app-phonology/src/main/java/net/calebscode/langforge/app/phonology/model/SyllablePatternCategoryMapModel.java
@@ -17,13 +17,16 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableMap;
 import net.calebscode.langforge.phonology.phoneme.Phoneme;
 import net.calebscode.langforge.phonology.syllable.SyllablePatternCategoryMap;
+import net.calebscode.langforge.phonology.syllable.SyllablePatternCompiler;
 
-public class SyllablePatternCategoryModel extends SyllablePatternCategoryMap {
+public class SyllablePatternCategoryMapModel extends SyllablePatternCategoryMap {
 
-	private MapProperty<Character, Set<Phoneme>> categoryMap;
-	private ReadOnlyMapWrapper<Character, Set<Phoneme>> categoryWrapper;
+	private final MapProperty<Character, Set<Phoneme>> categoryMap;
+	private final ReadOnlyMapWrapper<Character, Set<Phoneme>> categoryWrapper;
 
-	public SyllablePatternCategoryModel() {
+	private final SyllablePatternCompiler compiler;
+
+	public SyllablePatternCategoryMapModel() {
 		categoryMap = new SimpleMapProperty<>(observableHashMap());
 
 		categoryWrapper = new ReadOnlyMapWrapper<>();
@@ -34,10 +37,16 @@ public class SyllablePatternCategoryModel extends SyllablePatternCategoryMap {
 			}
 			return unmodifiableObservableMap(mapCopy);
 		}, categoryMap));
+
+		compiler = new SyllablePatternCompiler(this);
 	}
 
 	public ReadOnlyMapProperty<Character, Set<Phoneme>> categoryMapProperty() {
 		return categoryWrapper;
+	}
+
+	public SyllablePatternCompiler getCompiler() {
+		return compiler;
 	}
 
 	@Override
