+++ langforge-app-phonology/src/main/resources/fxml/SyllableManagementView.fxml
@@ -1,61 +1,84 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.Double?>
-
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.TextField?>
 
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
 
-<fx:root type="net.calebscode.langforge.app.phonology.view.SyllablePatternView" xmlns:fx="http://javafx.com/fxml">
+<fx:root type="net.calebscode.langforge.app.phonology.controller.SyllableManagementController" xmlns:fx="http://javafx.com/fxml">
 	<VBox style="-fx-padding: 5px" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
 		<GridPane VBox.vgrow="ALWAYS" hgap="5" vgap="2">
-		
-			<!-- Categories -->
-			<Label text="Category"
+						
+			<!-- Syllable Patterns -->
+			<Label text="Patterns"
 				GridPane.columnIndex="0"
+				GridPane.rowIndex="0" />
+		
+			<ListView fx:id="patternsList"
+				GridPane.columnIndex="0"
+				GridPane.rowIndex="1"
+				editable="false"
+				fixedCellSize="-1"
+				prefWidth="200"/>
+			
+			<HBox GridPane.columnIndex="0" GridPane.rowIndex="2">
+				<Button text="New Pattern" onAction="#createPattern" />
+				<Button text="Delete Pattern(s)" onAction="#deletePattern" />
+			</HBox>
+	
+			
+			<!-- Categories -->
+			<Label text="Categories"
+				GridPane.columnIndex="1"
 				GridPane.rowIndex="0" />
 				
 			<ListView fx:id="categoriesList"
-				GridPane.columnIndex="0"
+				GridPane.columnIndex="1"
 				GridPane.rowIndex="1"
 				editable="false" />
 			
-			<HBox GridPane.columnIndex="0" GridPane.rowIndex="2" >
+			<HBox GridPane.columnIndex="1" GridPane.rowIndex="2" >
 				<Button text="New" onAction="#createCategory" />
 				<Button text="Delete" onAction="#removeCategory" />
 			</HBox>
 			
+			
 			<!-- Phonemes for Category -->
 			<Label text="Phonemes for Category"
-				GridPane.columnIndex="1"
+				GridPane.columnIndex="2"
 				GridPane.rowIndex="0" />
 			
 			<ListView fx:id="phonemesList"
-				GridPane.columnIndex="1"
+				GridPane.columnIndex="2"
 				GridPane.rowIndex="1" />
 			
-			<HBox GridPane.columnIndex="1" GridPane.rowIndex="2" >
+			<HBox GridPane.columnIndex="2" GridPane.rowIndex="2" >
 				<Button text="Quick Add" />
 			</HBox>
 			
+			
 			<!-- Phoneme Movement Buttons -->
-			<VBox GridPane.columnIndex="2" GridPane.rowIndex="1">
+			<VBox GridPane.columnIndex="3" GridPane.rowIndex="1">
 				<Button text="&lt;" minWidth="25" onAction="#addToCategory" />
 				<Button text="&gt;" minWidth="25" onAction="#removeFromCategory" />
 			</VBox>
 			
+			
 			<!-- Phoneme Inventory -->
-			<Label text="Phoneme Inventory" GridPane.columnIndex="3" GridPane.rowIndex="0" />
+			<Label text="Phoneme Inventory"
+				GridPane.columnIndex="4"
+				GridPane.rowIndex="0" />
+			
 			<ListView fx:id="inventoryList"
-				GridPane.columnIndex="3"
+				GridPane.columnIndex="4"
 				GridPane.rowIndex="1"
 				editable="false" />
-			
+
 		</GridPane>
 	</VBox>
 </fx:root>
